+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Added
 -----
 
 * Show list of recent files on welcome screen
+* Keyboard shortcuts can now be configured via a settings file.
+  Go to "Settings -> Open Settings Folder" and edit KeyboardSettings.ini
 
 Fixed
 -----